verifier: move under training/renderer/, embed UI screenshots in README
The verifier validates renderers — colocating it under
training/renderer/verifier/ makes that ownership obvious and matches
how the cookbook organises related code (tests live next to what they
test, etc).

What moved:
  training/verifier/  →  training/renderer/verifier/

What changed:
- All Python imports retargeted to training.renderer.verifier.*.
- Workspace-root scripts (run.sh, triage.sh, run-bug-*.sh) updated to
  invoke `python -m training.renderer.verifier.*`.
- skills/verifier/SKILL.md, skills/renderer/SKILL.md, cookbook/README.md
  point at the new path.
- training/renderer/verifier/README.md now embeds two screenshots
  (config form + token stream) from the new images/ subfolder so
  visitors see the GUI before reading the long-form skill.
- Tests (test_verifier_probe.py, test_renderer_hf_parity.py) updated
  to the new module paths.

CLI invocations are now `python -m training.renderer.verifier render`
etc. — verbose, but it's the price of correct nesting. The shell
scripts hide this; only direct CLI users will notice.

9 passed + 1 xfail.
