+++ training/renderer/verifier/__main__.py
@@ -1,0 +1,6 @@
+"""Entry point: ``python -m training.renderer.verifier ...``."""
+
+from training.renderer.verifier.cli import main
+
+if __name__ == "__main__":
+    raise SystemExit(main())
