+++ training/renderer/verifier/utils/inspect_rules.py
