--- training/verifier/__main__.py
@@ -1,6 +1,0 @@
-"""Entry point: ``python -m training.verifier ...``."""
-
-from training.verifier.cli import main
-
-if __name__ == "__main__":
-    raise SystemExit(main())
